fix(auth): JWT iat as float (µs) so token revocation + same-second re-login both work
Previous fix stamped min_token_iat 1s in the future to outrun whole-second
iat collisions, but that over-revoked legitimate new tokens issued in the
following second too. Switching iat to float (RFC 7519 NumericDate allows it)
gives microsecond resolution: a re-login that happens 50ms after a revoke
will have iat > min_token_iat by ~50ms and pass.

+++ backend/app/auth/dependencies.py
@@ -36,9 +36,11 @@ async def get_current_user(
         raise HTTPException(status.HTTP_401_UNAUTHORIZED, detail="user_disabled")
 
     # Token revocation: reject any JWT issued before user.min_token_iat.
+    # Both sides are float seconds (µs precision) so same-second new logins
+    # after a revocation get a fractionally-larger iat and pass the strict <.
     if user.min_token_iat is not None:
         iat = payload.get("iat")
-        if iat is not None and int(iat) < int(user.min_token_iat.timestamp()):
+        if iat is not None and float(iat) < user.min_token_iat.timestamp():
             raise HTTPException(status.HTTP_401_UNAUTHORIZED, detail="token_revoked")
     return user
